Move files to sub-package digest
