+++ library/digest/src/commonMain/kotlin/org/kotlincrypto/core/digest/Digest.kt
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.core
+package org.kotlincrypto.core.digest
 
-import org.kotlincrypto.core.internal.DigestState
+import org.kotlincrypto.core.*
+import org.kotlincrypto.core.digest.internal.DigestState
 
 /**
  * Core abstraction for Message Digest implementations.
