+++ library/digest/src/commonTest/kotlin/org/kotlincrypto/core/digest/DigestUnitTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.core
+package org.kotlincrypto.core.digest
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
