+++ library/digest/src/commonTest/kotlin/org/kotlincrypto/core/digest/TestDigestException.kt
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.core
+package org.kotlincrypto.core.digest
 
+import org.kotlincrypto.core.Updatable
 import kotlin.test.Test
 import kotlin.test.fail
