+++ library/digest/src/jvmTest/kotlin/org/kotlincrypto/core/digest/JvmDigestUnitTest.kt
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package org.kotlincrypto.core
+package org.kotlincrypto.core.digest
 
-import org.kotlincrypto.core.internal.DigestState
+import org.kotlincrypto.core.InternalKotlinCryptoApi
+import org.kotlincrypto.core.digest.internal.DigestState
 import java.lang.AssertionError
 import java.security.MessageDigest
 import kotlin.random.Random
